+++ src/widgets/helper/ChannelView.hpp
@@ -193,6 +193,12 @@ public:
      */
     bool mayContainMessage(const MessagePtr &message);
 
+    void updateColorTheme();
+    void setColorVisitor(
+        const std::function<void(MessageColors &, Theme *)> &visitor);
+
+    Scrollbar *scrollbar();
+
     pajlada::Signals::Signal<QMouseEvent *> mouseDown;
     pajlada::Signals::NoArgSignal selectionChanged;
     pajlada::Signals::Signal<HighlightState> tabHighlightRequested;
@@ -410,6 +416,8 @@ private:
 
     MessageColors messageColors_;
     MessagePreferences messagePreferences_;
+
+    std::function<void(MessageColors &, Theme *)> colorVisitor_;
 
     void scrollUpdateRequested();